+++ packages/widget_toolkit_pin/lib/src/lib_pin_code_with_biometrics/blocs/pin_code_bloc.rxb.g.dart
@@ -19,60 +19,29 @@ abstract class $PinCodeBloc extends RxBlocBase
     implements PinCodeBlocEvents, PinCodeBlocStates, PinCodeBlocType {
   final _compositeSubscription = CompositeSubscription();
 
-  /// Тhe [Subject] where events sink to by calling [getPinLength]
-  final _$getPinLengthEvent = PublishSubject<int?>();
+  /// Тhe [Subject] where events sink to by calling [addDigit]
+  final _$addDigitEvent = PublishSubject<String>();
 
-  /// Тhe [Subject] where events sink to by calling [checkBiometricsEnabled]
-  final _$checkBiometricsEnabledEvent = BehaviorSubject<void>();
+  /// Тhe [Subject] where events sink to by calling [deleteDigit]
+  final _$deleteDigitEvent = PublishSubject<void>();
 
-  /// Тhe [Subject] where events sink to by calling [checkAvailableBiometrics]
-  final _$checkAvailableBiometricsEvent = BehaviorSubject<void>();
+  /// Тhe [Subject] where events sink to by calling [biometricsButtonPressed]
+  final _$biometricsButtonPressedEvent = PublishSubject<void>();
 
-  /// Тhe [Subject] where events sink to by calling [checkPinCodeInStorage]
-  final _$checkPinCodeInStorageEvent = BehaviorSubject<void>();
+  /// Тhe [Subject] where events sink to by calling [checkIfPinIsStored]
+  final _$checkIfPinIsStoredEvent = PublishSubject<void>();
 
-  /// Тhe [Subject] where events sink to by calling [requestBiometricAuth]
-  final _$requestBiometricAuthEvent = PublishSubject<String>();
+  /// The state of [digitsCount] implemented in [_mapToDigitsCountState]
+  late final Stream<int> _digitsCountState = _mapToDigitsCountState();
 
-  /// Тhe [Subject] where events sink to by calling [autoSubmit]
-  final _$autoSubmitEvent = PublishSubject<String>();
+  /// The state of [showBiometricsButton] implemented in
+  /// [_mapToShowBiometricsButtonState]
+  late final ConnectableStream<bool> _showBiometricsButtonState =
+      _mapToShowBiometricsButtonState();
 
-  /// Тhe [Subject] where events sink to by calling [setBiometrics]
-  final _$setBiometricsEvent =
-      PublishSubject<({bool enabled, String localizedReason})>();
-
-  /// The state of [areBiometricsEnabled] implemented in
-  /// [_mapToAreBiometricsEnabledState]
-  late final Stream<bool> _areBiometricsEnabledState =
-      _mapToAreBiometricsEnabledState();
-
-  /// The state of [isPinCodeVerified] implemented in
-  /// [_mapToIsPinCodeVerifiedState]
-  late final ConnectableStream<bool> _isPinCodeVerifiedState =
-      _mapToIsPinCodeVerifiedState();
-
-  /// The state of [availableBiometrics] implemented in
-  /// [_mapToAvailableBiometricsState]
-  late final ConnectableStream<List<BiometricsAuthType>>
-      _availableBiometricsState = _mapToAvailableBiometricsState();
-
-  /// The state of [isAuthenticatedWithBiometrics] implemented in
-  /// [_mapToIsAuthenticatedWithBiometricsState]
-  late final ConnectableStream<bool> _isAuthenticatedWithBiometricsState =
-      _mapToIsAuthenticatedWithBiometricsState();
-
-  /// The state of [biometricsDialog] implemented in
-  /// [_mapToBiometricsDialogState]
-  late final Stream<BiometricsMessage?> _biometricsDialogState =
-      _mapToBiometricsDialogState();
-
-  /// The state of [isPinCodeInSecureStorage] implemented in
-  /// [_mapToIsPinCodeInSecureStorageState]
-  late final ConnectableStream<bool> _isPinCodeInSecureStorageState =
-      _mapToIsPinCodeInSecureStorageState();
-
-  /// The state of [pinLength] implemented in [_mapToPinLengthState]
-  late final ConnectableStream<int> _pinLengthState = _mapToPinLengthState();
+  /// The state of [authenticated] implemented in [_mapToAuthenticatedState]
+  late final ConnectableStream<void> _authenticatedState =
+      _mapToAuthenticatedState();
 
   /// The state of [isLoading] implemented in [_mapToIsLoadingState]
   late final Stream<bool> _isLoadingState = _mapToIsLoadingState();
@@ -81,57 +50,26 @@ abstract class $PinCodeBloc extends RxBlocBase
   late final Stream<ErrorModel> _errorsState = _mapToErrorsState();
 
   @override
-  void getPinLength(int? length) => _$getPinLengthEvent.add(length);
+  void addDigit(String digit) => _$addDigitEvent.add(digit);
 
   @override
-  void checkBiometricsEnabled() => _$checkBiometricsEnabledEvent.add(null);
+  void deleteDigit() => _$deleteDigitEvent.add(null);
 
   @override
-  void checkAvailableBiometrics() => _$checkAvailableBiometricsEvent.add(null);
+  void biometricsButtonPressed() => _$biometricsButtonPressedEvent.add(null);
 
   @override
-  void checkPinCodeInStorage() => _$checkPinCodeInStorageEvent.add(null);
+  void checkIfPinIsStored() => _$checkIfPinIsStoredEvent.add(null);
 
   @override
-  void requestBiometricAuth(String localizedMessage) =>
-      _$requestBiometricAuthEvent.add(localizedMessage);
+  Stream<int> get digitsCount => _digitsCountState;
 
   @override
-  void autoSubmit(String pin) => _$autoSubmitEvent.add(pin);
+  ConnectableStream<bool> get showBiometricsButton =>
+      _showBiometricsButtonState;
 
   @override
-  void setBiometrics(
-    bool enabled,
-    String localizedReason,
-  ) =>
-      _$setBiometricsEvent.add((
-        enabled: enabled,
-        localizedReason: localizedReason,
-      ));
-
-  @override
-  Stream<bool> get areBiometricsEnabled => _areBiometricsEnabledState;
-
-  @override
-  ConnectableStream<bool> get isPinCodeVerified => _isPinCodeVerifiedState;
-
-  @override
-  ConnectableStream<List<BiometricsAuthType>> get availableBiometrics =>
-      _availableBiometricsState;
-
-  @override
-  ConnectableStream<bool> get isAuthenticatedWithBiometrics =>
-      _isAuthenticatedWithBiometricsState;
-
-  @override
-  Stream<BiometricsMessage?> get biometricsDialog => _biometricsDialogState;
-
-  @override
-  ConnectableStream<bool> get isPinCodeInSecureStorage =>
-      _isPinCodeInSecureStorageState;
-
-  @override
-  ConnectableStream<int> get pinLength => _pinLengthState;
+  ConnectableStream<void> get authenticated => _authenticatedState;
 
   @override
   Stream<bool> get isLoading => _isLoadingState;
@@ -139,19 +77,11 @@ abstract class $PinCodeBloc extends RxBlocBase
   @override
   Stream<ErrorModel> get errors => _errorsState;
 
-  Stream<bool> _mapToAreBiometricsEnabledState();
+  Stream<int> _mapToDigitsCountState();
 
-  ConnectableStream<bool> _mapToIsPinCodeVerifiedState();
+  ConnectableStream<bool> _mapToShowBiometricsButtonState();
 
-  ConnectableStream<List<BiometricsAuthType>> _mapToAvailableBiometricsState();
-
-  ConnectableStream<bool> _mapToIsAuthenticatedWithBiometricsState();
-
-  Stream<BiometricsMessage?> _mapToBiometricsDialogState();
-
-  ConnectableStream<bool> _mapToIsPinCodeInSecureStorageState();
-
-  ConnectableStream<int> _mapToPinLengthState();
+  ConnectableStream<void> _mapToAuthenticatedState();
 
   Stream<bool> _mapToIsLoadingState();
 
@@ -165,17 +95,11 @@ abstract class $PinCodeBloc extends RxBlocBase
 
   @override
   void dispose() {
-    _$getPinLengthEvent.close();
-    _$checkBiometricsEnabledEvent.close();
-    _$checkAvailableBiometricsEvent.close();
-    _$checkPinCodeInStorageEvent.close();
-    _$requestBiometricAuthEvent.close();
-    _$autoSubmitEvent.close();
-    _$setBiometricsEvent.close();
+    _$addDigitEvent.close();
+    _$deleteDigitEvent.close();
+    _$biometricsButtonPressedEvent.close();
+    _$checkIfPinIsStoredEvent.close();
     _compositeSubscription.dispose();
     super.dispose();
   }
 }
-
-// ignore: unused_element
-typedef _SetBiometricsEventArgs = ({bool enabled, String localizedReason});