+++ packages/widget_toolkit_pin/lib/src/lib_pin_code_with_biometrics/models/error_wrong_pin.dart
@@ -1,7 +1,7 @@
 import 'package:widget_toolkit/models.dart';
 
 /// This error can be thrown from the service layer, when the user types a wrong
-/// pin code. When the [PinCodeComponent] widget receives this error a shake
+/// pin code. When the PinCodeComponent widget receives this error a shake
 /// animation will be triggered on the masked pin code keys, then they disappear
 /// and the text from the [errorMessage] will be displayed on the screen.
 class ErrorWrongPin extends ErrorModel {