+++ packages/widget_toolkit_pin/lib/src/lib_pin_code_with_biometrics/services/pin_code_service.dart
@@ -1,13 +1,14 @@
 abstract class PinCodeService {
-  /// Reads from device internal storage and returns the stored pin code to the bloc
+  /// Reads from device internal storage and
+  /// returns the stored pin code to the bloc
   Future<bool> isPinCodeInSecureStorage();
 
-  /// Receives the encrypted pinCode from user input. Returns whether it was
-  /// verified by the server
+  /// Receives the encrypted pinCode from user input.
+  /// Returns whether it was verified by the server
   Future<bool> verifyPinCode(String pinCode);
 
-  /// Returns the correct length from the backend, the length should be less
-  /// than 10 digits
+  /// Returns the correct length from the backend,
+  /// the length should be less than 10 digits
   Future<int> getPinLength();
 
   /// Encrypts the string and stores it in the device secure storage. It has a